Added default file to load from and separated code contents. removed old
