+++ deimain.c
@@ -1,81 +1,53 @@
 
-/* All 3 for printing backtrace on crash */
-#include <execinfo.h>
-#include <signal.h>
-#include <unistd.h>
+
+#include <stdarg.h> /* needed for yyerror */
 
 #include "dei.tab.h"
 
 #include "deimain.h"
-#include "struct.h"
-#include "evaluation.h"
-
-FILE *logger;
-
-void handler(int sig) {
-  void *array[10];
-  size_t size;
-
-  // get void*'s for all entries on the stack
-  size = backtrace(array, 10);
-
-  // print out all the frames to stderr
-  fprintf(stderr, "Error: signal %d:\n", sig);
-  backtrace_symbols_fd(array, size, STDERR_FILENO);
-  exit(1);
-}
 
 int main(int argc, char **argv){
-  signal(SIGSEGV, handler);   // install crash handler -> displays backtrace to stderr
+	char *stdsave = "config.dei";
 
-  #ifdef DEBUG
-  #ifdef YYDEBUG
-  yydebug = 1;
-  #endif
+	FILE *f = fopen(stdsave, "r");
+	if (f){
+		printf("found save!\n");
+		yyin = f;
+		yyparse();
+		fclose(yyin);
+		printf("\n");
+	}
 
-  logger = fopen("dei.log", "a+"); // a+ (create + append) option will allow appending which is useful in a log file
-  if (logger == NULL) { perror("Failed: "); return 1; }
+	int i;
+	for(i = 1; i < argc; i++) {
+		f = fopen(argv[i], "r");
 
-  time_t rawtime;
-  struct tm *timeinfo;
-  time(&rawtime);
-  timeinfo = localtime(&rawtime);
+		if(!f) {
+			perror(argv[i]);
+			return (1);
+		}
 
-  fprintf(logger, "=== %s  ===\n", asctime(timeinfo));
-
-
-  printf("value bytes: %lu\n", sizeof(struct value));
-  printf("symbol bytes: %lu\n", sizeof(struct symbol));
-  printf("die bytes: %lu\n", sizeof(struct die));
-  printf("roll bytes: %lu\n", sizeof(struct roll));
-  printf("result bytes: %lu\n", sizeof(struct result));
-  printf("\n");
-  printf("ast bytes: %lu\n", sizeof(struct ast));
-  printf("natdie bytes: %lu\n", sizeof(struct natdie));
-  printf("setdie bytes: %lu\n", sizeof(struct setdie));
-  printf("natint bytes: %lu\n", sizeof(struct natint));
-  printf("setres bytes: %lu\n", sizeof(struct setres));
-  printf("funcall bytes: %lu\n", sizeof(struct funcall));
-  printf("symcall bytes: %lu\n", sizeof(struct symcall));
-  printf("symasgn bytes: %lu\n\n", sizeof(struct symasgn));
-
-  #endif
-
-  /* seed rng with current time */
-  srand(time(NULL));
-
-  if (argc > 1){
-    if (!(yyin = fopen(argv[1], "r"))){
-      perror(argv[1]);
-      return 1;
-    }
-  }else{
-    printf("> ");
+		printf("new file!\n");
+		yyin = f;
+		yylineno = 1;
+		yyrestart(yyin);
+		yyparse();
+		fclose(yyin);
+		printf("\n");
   }
 
-  int end = yyparse();
+	printf("stdin!\n> ");
+	yyin = stdin;
+	yylineno = 1;
+	yyrestart(yyin);
+  return yyparse();
+}
 
-  //freesymboltable();
+void yyerror(char *s, ...){
+  va_list ap;
+  va_start(ap, s);
 
-  return end;
+  fprintf(stderr, "%d: error : ", yylineno);
+  vfprintf(stderr, s, ap);
+  fprintf(stderr, "\n");
 }
